288 feat 조건에 따른 월간 파트타이머 조회 및 수정 api 개발

+++ app/cruds/labor_management/dto/part_timers_response_dto.py
@@ -8,8 +8,8 @@ class PartTimerSummaryResponseDTO(BaseModel):
     user_name: str
     part_name: str
     work_days: int
-    hospital_work_hours: int
-    holiday_work_hours: int
+    hospital_work_hours: float
+    holiday_work_hours: float
     total_work_hours: float
     total_wage: float
     phone_number: str
